+++ src/client/extension.ts
@@ -63,6 +63,7 @@ import { IDebugSessionEventHandlers } from './debugger/extension/hooks/types';
 import { registerTypes as debugConfigurationRegisterTypes } from './debugger/extension/serviceRegistry';
 import { IDebugConfigurationService, IDebuggerBanner } from './debugger/extension/types';
 import { registerTypes as formattersRegisterTypes } from './formatters/serviceRegistry';
+import { IInterpreterAutoSeletionService } from './interpreter/autoSelection/types';
 import { IInterpreterSelector } from './interpreter/configuration/types';
 import {
     ICondaService,
@@ -71,7 +72,6 @@ import {
     InterpreterLocatorProgressHandler,
     PythonInterpreter
 } from './interpreter/contracts';
-import { IInterpreterAutoSeletionService } from './interpreter/interpreterSelection/types';
 import { registerTypes as interpretersRegisterTypes } from './interpreter/serviceRegistry';
 import { ServiceContainer } from './ioc/container';
 import { ServiceManager } from './ioc/serviceManager';
