+++ src/client/interpreter/autoSelection/stratergies/currentPath.ts
