+++ src/client/interpreter/autoSelection/stratergies/windowsRegistry.ts
