+++ src/test/datascience/execution.unit.test.ts
@@ -33,8 +33,8 @@ import { Architecture } from '../../client/common/utils/platform';
 import { EXTENSION_ROOT_DIR } from '../../client/constants';
 import { JupyterExecution } from '../../client/datascience/jupyterExecution';
 import { ICell, IConnection, IJupyterKernelSpec, INotebookServer, InterruptResult } from '../../client/datascience/types';
+import { InterpreterAutoSeletionService } from '../../client/interpreter/autoSelection';
 import { InterpreterType, PythonInterpreter } from '../../client/interpreter/contracts';
-import { InterpreterAutoSeletionService } from '../../client/interpreter/interpreterSelection';
 import { InterpreterService } from '../../client/interpreter/interpreterService';
 import { CondaService } from '../../client/interpreter/locators/services/condaService';
 import { KnownSearchPathsForInterpreters } from '../../client/interpreter/locators/services/KnownPathsService';
